backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
groq==0.4.2
anthropic==0.21.3
python-dotenv==1.0.1
pydantic==2.6.4
resend==2.1.0
